experiment of OutPutParsers and prompt templates

+++ src/main/java/dev/chol/anyquizai/domain/Quiz.java
@@ -26,7 +26,7 @@ public class Quiz {
     private String title;
 
     @Column
-    private Long totalQuestions;
+    private Integer totalQuestions;
 
     @Column
     @Enumerated(EnumType.STRING)
